refactor menu bar as a reuseable component
